requirements.txt
numpy
pylatex
#python-nmap
pydot
psutil
ipwhois
#whois
python-whois
